+++ examples/format_customized.py
@@ -1,12 +1,16 @@
-from traceback_with_variables import Format, ColorSchemes, is_ipython_global, fmt
+from traceback_with_variables import Format, ColorSchemes, is_ipython_global, fmt, hide, skip
 
 
-# approach 1
+# approach 1 for simples usage
 
 fmt.max_value_str_len = 10000
 
+# show all vars, no hides, no skips:
 
-# approach 2
+fmt.custom_var_printers = []
+
+
+# approach 2 for explicit format passing
 
 fmt2 = Format(
     before=3,
@@ -20,17 +24,17 @@ fmt2 = Format(
     skip_files_except=['my_project', 'site-packages'],
     brief_files_except='my_project',
     custom_var_printers=[  # first matching is used
-        ('password', lambda v: '...hidden...'),  # by name, print const str
+        ('password', hide),  # by name, print const str
         (list, lambda v: f'list{v}'),  # by type, print fancy str
-        (lambda name, type_, filename, is_global: is_global, lambda v: None),  # custom filter, skip printing
+        (lambda name, type_, filename, is_global: is_global, skip),  # custom filter, skip printing
         (is_ipython_global, lambda v: None),  # same, handy for Jupyter
         (['secret', dict, (lambda name, *_: 'asd' in name)], lambda v: '???'),  # by different things, print const str
     ]
 )
-# print_exc(fmt=fmt2)
+print_exc(fmt=fmt2)
 
 
-# approach 3
+# approach 3 for explicit format passing, less code
 
 fmt3 = defaults.replace(max_value_str_len=10000)
-# print_exc(fmt=fmt3)
+print_exc(fmt=fmt3)